requirements.txt
# Deep Learning Capstone - SSL Project
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data handling
numpy>=1.24.0
Pillow>=9.0.0
datasets>=2.14.0  # HuggingFace datasets
huggingface-hub>=0.16.0
# learn2learn>=0.2.0  # Required for Mini-ImageNet download

# Augmentations
albumentations>=1.3.0

# Evaluation & ML utilities
scikit-learn>=1.3.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Logging & Monitoring
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Optional: Distributed training
# accelerate>=0.21.0
